+++ deploy/mc-deploy/CHANGELOG.md
@@ -1,5 +1,8 @@
 ALSO: UPDATE version in pyproject.toml, add/push mc-deploy-X.Y.Z tag to github
 
+* 0.5.0: initial work on docker.py: pull AIRTABLE/TZ settings into base
+	add port_bias (for indexer), make INST_FLAVOR map values a tuple
+
 * 0.4.2: py.typed committed!
 * 0.4.1: py.typed
 * 0.4.0: DokkuDBDjangoDeploy